Added cname_resolve
cname_resolve is a shell script that takes a list of subdomains and a root domain, appends the root domain to the subdomains, and checks the cnames of each subdomain to see if they resolve or not. Can be used for general recon, or to look for potential subdomain takeovers. Inspired by TomNomNom